Fixing ANSI codes for dim pipe colors
The `0` in SGR resets all color info, so we don't want to use that in the ANSI to pipe-code translation. This was causing a problem for some strings using a background color change immediately before a "dim" foreground color change. For example: `|21|00test string` was resulting in black foreground on black background, when the intent was black foreground on a magenta background.

`2` sets standard (dim) intensity, so using that instead for pipe color codes 00-07.